Refactor motion variants into a dedicated file for better organization; update imports in Personal and Project components to use the new motion module.
