+++ app/page.tsx
@@ -7,25 +7,7 @@ import { Blogs } from '@/components/blogs';
 import { Contact } from '@/components/contact';
 import { Education } from '@/components/education';
 import { Skills } from '@/components/skills';
-
-export const VARIANTS_CONTAINER = {
-  hidden: { opacity: 0 },
-  visible: {
-    opacity: 1,
-    transition: {
-      staggerChildren: 0.15,
-    },
-  },
-};
-
-export const VARIANTS_SECTION = {
-  hidden: { opacity: 0, y: 20, filter: 'blur(8px)' },
-  visible: { opacity: 1, y: 0, filter: 'blur(0px)' },
-};
-
-export const TRANSITION_SECTION = {
-  duration: 0.3,
-};
+import { VARIANTS_CONTAINER, VARIANTS_SECTION, TRANSITION_SECTION } from '@/components/ui/motion';
 
 export default function Personal() {
   return (
